Reduce ram usage by a lot, demo code in adapter working

+++ lg_ssd.h
@@ -1,13 +1,13 @@
 #include <avr/io.h>
 
-
 void reset_SSDs();
 void spin_SSDs();
-void update_ssd0(int ssd_val);
-void update_ssd1(int ssd_val);
-void update_LED(int status);
-void update_relay(int status);
+void update_ssd0(uint8_t ssd_val);
+void update_ssd1(uint8_t ssd_val);
+void update_LED(uint8_t status);
+void update_relay(uint8_t status);
 void init_SSDs();
 
 void SSD0_shiftin(bool data);
 void SSD1_shiftin(bool data);
+
